requirements.txt
arrow==0.5.4
chai==1.1.0
coverage==3.7.1
docker-py==1.2.2
flake8==2.4.1
humanize==0.5.1
mccabe==0.3
nose==1.3.7
pep8==1.5.7
pyflakes==0.8.1
python-dateutil==2.4.2
requests==2.7.0
six==1.9.0
wheel==0.24.0
